+++ ChangeLog.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 
 ## Version 3.6
+- FIX : Title lines broken on PDF for documents whose lines use the `desc` field instead of the `label` field (such as
+        supplier orders and invoices) *22/11/2021* - 3.6.10
+- FIX : checkbox to add a subtotal title block per order on invoices using the "Bill orders" feature was broken by
+        core changes in Dolibarr *12/11/2021* - 3.6.9
 - FIX : Missing columns in invoice creation page (from order) *10/06/2022* - 3.6.8
 - FIX : addition of a conf allowing to add the subtotal line or not when creating an expedition from an order *12/07/2021* - 3.6.7 
 - FIX : Clone icon compatibility *08/06/2021* - 3.6.6
@@ -27,7 +31,7 @@ All notable changes to this project will be documented in this file.
   sur une nouvelle release - 3.5
 
 ## Version 3.5
-
+- FIX: invoice creation: title/subtotal/free text lines coming from shipments or deliveries not imported with special code (MDLL) - *17/11/2021* - 3.5.8
 - FIX : Text or title line break PDF *15/04/2021* - 3.5.7
 - NEW Ajouter les lignes 'Titre' , 'total' , 'libre' aux generation d'expedition de commandes expédiables (il faudra supprimer les lignes de sous-total à la main si le besoin s'en fait sentir ) [2021-02-03]